[FLINK-27998][table] Address feedback for Ser/de for three-value Search operator
